Create Bootstrap 4 nav toggle collapse automatically when clicking away
JavaScript to make the bootstrap 4 toggle/burger menu collapse automatically if the user opens the menu and decides not to click a menu link, but wants to continue clicking and scrolling on the current page. 
(Bootstrap's default collapsible burger menu has to be collapsed manually by clicking the icon if the user doesnt click a link - the user cannot seemlessly return to scrolling down the current page).